refactor(evaluation): declare completion-window envs in config, split the eval jobs table module
The O2_EVAL_TRACE/SESSION_* completion-window envs move from raw
std::env reads in infra into the enterprise LlmEvaluationConfig where
every other eval env lives; startup wiring injects the values into infra
via init_completion_window_defaults (same registration pattern as the
search executor), and the built-in constants remain the fallback when
the wiring never runs.

online_eval_jobs.rs also splits into submodules — completion_window.rs
(defaults, guard rails, trace/session configs) and span_selector.rs —
with re-exports keeping every external path unchanged.
